Extract schema transitions under delegated namespace
